requirements.txt
beautifulsoup4==4.10.0
loguru==0.5.3
requests==2.26.0
selenium==4.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.32.0
